pyproject.toml
[project]
name = "swhid-harness"
version = "2.0.0"
description = "Technology-neutral testing harness for SWHID implementations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0"}
authors = [
    { name = "Software Heritage", email = "contact@softwareheritage.org" }
]
dependencies = [
    "PyYAML==6.0.2",
    "pydantic==2.12.0",
    "psutil==6.0.0",
    "dulwich==0.21.7",
    "pygit2==1.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-cov==5.0.0", 
    "pytest-mock==3.14.0",
    "ruff==0.6.9",
    "mypy==1.11.2"
]

[project.scripts]
swhid-harness = "harness.harness:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["harness*"]

[tool.setuptools.package-data]
harness = ["py.typed"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203", "E266", "E501"]

[tool.pytest.ini_options]
addopts = "--tb=short -q"
testpaths = ["tests"]
